backend/requirements.txt
flask==3.0.3
python-telegram-bot==21.4
python-dotenv==1.0.1